+++ src/app/(dashboard)/account/hooks/delete/use-delete.ts
@@ -1,0 +1,11 @@
+import { useState } from 'react';
+
+export default function useAccountDelete() {
+  const [dialogOpen, setDialogOpen] = useState(false);
+
+  const handleDialogOpen = () => {
+    setDialogOpen((prevState) => !prevState);
+  };
+
+  return { dialogOpen, handleDialogOpen };
+}