+++ src/app/(dashboard)/account/hooks/name/use-name.ts
@@ -1,0 +1,11 @@
+import { useState } from 'react';
+
+export default function useAccountName() {
+  const [formShown, setFormShown] = useState(false);
+
+  const handleFormShown = () => {
+    setFormShown((prevState) => !prevState);
+  };
+
+  return { formShown, handleFormShown };
+}